Overhaul fmg-core state-related code
Currently, fmg-core assumes that the consumer will extend the State
class, and implement its own gameAttributes function.

In fact, this doesn't work, since a generic wallet doesn't know the shape
of the game attributes for an arbitrary application, and the application
shouldn't have to worry about framework attributes.

Moreover, the wallet won't want to instantiate objects of the State
class, since it will want to handle vanilla objects that will be stored
in the redux store.

Therefore, the state.ts module exports the State interface specifying
the shape of a state object, and some helper functions, useful for
interacting with the blockchain